pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agraph-python"
dynamic = ["version"]
description = "AllegroGraph Python client"
readme = "README.rst"
authors = [
  {name = "Franz Inc.", email = "support@franz.com"}
]
license = {file = "LICENSE"}
requires-python = ">=3.7"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Topic :: Database",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "requests[socks]==2.25.0",
  "iso8601==0.1.12",
  "packaging~=23.2.0",
  "simplejson~=3.19.0",
  "pandas~=0.23.0",
]

[project.optional-dependencies]
curl = ["pycurl>=7.19.0,<8.0"]
dev = [
  "ipython~=7.34.0",
  "pytest~=7.0.0",
  "pytest-timeout~=2.2.0",
  "pytest-mock~=3.11.0",
  "pytest-isort~=3.1.0",
  "pytest-black~=0.3.0",
  "isort~=5.11.0",
  "black~=23.3.0",
  "nose==1.3.7",
  "sphinxcontrib-jupyter==0.1.dev20171024", # for building docs
]

[project.urls]
homepage = "https://github.com/franzinc/agraph-python"
documentation = "https://franz.com/agraph/support/documentation/current/python/"
repository = "https://github.com/franzinc/agraph-python"
changelog = "https://franz.com/agraph/support/documentation/current/python/changes.html"

[tool.setuptools.dynamic]
version = { attr = "franz.__version__" }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
minversion = "7.0.0"
addopts=[
  "--timeout=300",
  "--junitxml=report.xml",
  "--color=auto",
  "--isort",
  "--black"
]
testpaths = [
  "src/franz/openrdf/tests/",
  "src/franz/miniclient/",
]
python_files = ["*test*.py"]

[tool.pyright]
include = ["src"]
exclude = [
  "**/node_modules",
  "**/_pycache__",
  ".venv/",
]
root = "src"
reportMissingImports = true
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
pythonPlatform = "All"
verboseOutput = true

[tool.isort]
atomic = true
profile = "black"
src_paths = ["src"]
extend_skip = ["stress"]
filter_files = true
line_length = 88

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
(
    ^/stress/
)
'''
